Added filter for shakespeare
This filter reads the shakespeare text, drops lines that are just CAPITALS or
empty (leaving lines like 'Exeunt FAIRIES\n'), and if it drops a line inserts an
end-of-sentence token.  It also assumes that periods are end-of-sentence tokens,
since I couldn't find any that weren't on a cursory inspection.

Other punctuation is left unmarred.  The words are output as a singe list